refactor: update import paths for FilterContent and SnackBarState
